+++ docs/source/changelog.rst
@@ -2,6 +2,16 @@ Changelog
 =========
 
 
+v26.1.11
+--------
+
+**Added Changes**
+
+- Receive all invite-related gateway events.
+- Added :class:`~discord.events.InviteDeleteEvent` event.
+- Added :class:`~discord.events.InviteCreateEvent` event.
+
+
 v26.1.10
 --------